Condition regulations

+++ HastaneYönetim/HastaneYönetim/SekreterBransAyarForm.cs
@@ -19,27 +19,37 @@ namespace HastaneYönetim
 
         SqlBaglantisi bgl = new SqlBaglantisi();
 
-        private void SekreterBransAyarForm_Load(object sender, EventArgs e)
+        void BransCek()
         {
-            //bransları datagride çekme
             DataTable dt1 = new DataTable();
             SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Branslar", bgl.baglanti());
             da1.Fill(dt1);
             dataGridView1.DataSource = dt1;
+
+        }
+        private void SekreterBransAyarForm_Load(object sender, EventArgs e)
+        {
+            //bransları datagride çekme
+            BransCek();
         }
 
         private void BtnEkle_Click(object sender, EventArgs e)
         {
-            SqlCommand BransEkle=new SqlCommand("insert into Tbl_Branslar (BransAd) values (@b1)",bgl.baglanti());
-            BransEkle.Parameters.AddWithValue("@b1",TxtBransAd.Text);
-            BransEkle.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Kayıt başarıyla eklendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //branşları tekrar çekme
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Branslar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+           
+            if (TxtBransAd.Text!="")
+            {
+                SqlCommand BransEkle = new SqlCommand("insert into Tbl_Branslar (BransAd) values (@b1)", bgl.baglanti());
+                BransEkle.Parameters.AddWithValue("@b1", TxtBransAd.Text);
+                BransEkle.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Kayıt başarıyla eklendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //branşları tekrar çekme
+                BransCek();
+            }
+            else
+            {
+                MessageBox.Show("Lütfen eklemek istediğiniz branşı giriniz.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -52,31 +62,46 @@ namespace HastaneYönetim
 
         private void BtnGuncelle_Click(object sender, EventArgs e)
         {
-            SqlCommand BransGuncelle = new SqlCommand("Update Tbl_Branslar Set BransAd=@b1 Where Bransid=@b2", bgl.baglanti());
-            BransGuncelle.Parameters.AddWithValue("@b1", TxtBransAd.Text);
-            BransGuncelle.Parameters.AddWithValue("@b2", TxtBransid.Text);
-            BransGuncelle.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Güncelleme başarılı.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //branşları tekrar çekme
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Branslar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+           
+            if (TxtBransAd.Text!="" && TxtBransid.Text!="")
+            {
+                SqlCommand BransGuncelle = new SqlCommand("Update Tbl_Branslar Set BransAd=@b1 Where Bransid=@b2", bgl.baglanti());
+                BransGuncelle.Parameters.AddWithValue("@b1", TxtBransAd.Text);
+                BransGuncelle.Parameters.AddWithValue("@b2", TxtBransid.Text);
+                BransGuncelle.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Güncelleme başarılı", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //branşları tekrar çekme
+                BransCek();
+            }
+            else
+            {
+                MessageBox.Show("Lütfen güncellemek istediğiniz branşı seçiniz", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void BtnSil_Click(object sender, EventArgs e)
         {
-            SqlCommand BransSil = new SqlCommand("Delete From Tbl_Branslar Where Bransid=@b1", bgl.baglanti());
-            BransSil.Parameters.AddWithValue("@b1", TxtBransid.Text);
-            BransSil.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Kayıt başarıyla silindi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //doktorları tekrar gridview çek
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Branslar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+            if (TxtBransAd.Text != "" && TxtBransid.Text != "")
+            {
+                SqlCommand BransSil = new SqlCommand("Delete From Tbl_Branslar Where Bransid=@b1", bgl.baglanti());
+                BransSil.Parameters.AddWithValue("@b1", TxtBransid.Text);
+                BransSil.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Kayıt başarıyla silindi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //bransları tekrar gridview çek
+                BransCek();
+            }
+            else
+            {
+                MessageBox.Show("Lütfen silmek istediğiniz branşı seçiniz", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            TxtBransAd.Text = "";
+            TxtBransid.Text = "";
         }
     }
 }
