+++ HastaneYönetim/HastaneYönetim/SekreterDoktorAyarForm.cs
@@ -19,16 +19,19 @@ namespace HastaneYönetim
         }
         SqlBaglantisi bgl = new SqlBaglantisi();
 
-
+        void DoktorCek()
+        {
+            DataTable dt1 = new DataTable();
+            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Doktorlar", bgl.baglanti());
+            da1.Fill(dt1);
+            dataGridView1.DataSource = dt1;
+        }
 
         private void SekreterDoktorAyarForm_Load(object sender, EventArgs e)
         {
             //doktorları datagridviewe çek
 
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Doktorlar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+            DoktorCek();
 
 
             //branşları comboboxa çekme
@@ -42,21 +45,25 @@ namespace HastaneYönetim
         }
 
         private void BtnEkle_Click(object sender, EventArgs e)
-        {
-            SqlCommand DoktorEkle= new SqlCommand("insert into Tbl_Doktorlar (DoktorAd,DoktorSoyad,DoktorBrans,DoktorTC,DoktorSifre) values (@d1,@d2,@d3,@d4,@d5)",bgl.baglanti());
-            DoktorEkle.Parameters.AddWithValue("@d1", TxtAd.Text);
-            DoktorEkle.Parameters.AddWithValue("@d2", TxtSoyad.Text);
-            DoktorEkle.Parameters.AddWithValue("@d3", CmbBrans.Text);
-            DoktorEkle.Parameters.AddWithValue("@d4", MskTc.Text);
-            DoktorEkle.Parameters.AddWithValue("@d5", TxtSifre.Text);
-            DoktorEkle.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Doktor Eklemesi Yapıldı", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //doktorları tekrar gridview çek
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Doktorlar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+        {  
+            if (MskTc.Text!="" && MskTc.MaskCompleted && TxtAd.Text!="" && TxtSoyad.Text!="" && CmbBrans.Text!="" && TxtSifre.Text!="")
+            {
+                SqlCommand DoktorEkle = new SqlCommand("insert into Tbl_Doktorlar (DoktorAd,DoktorSoyad,DoktorBrans,DoktorTC,DoktorSifre) values (@d1,@d2,@d3,@d4,@d5)", bgl.baglanti());
+                DoktorEkle.Parameters.AddWithValue("@d1", TxtAd.Text);
+                DoktorEkle.Parameters.AddWithValue("@d2", TxtSoyad.Text);
+                DoktorEkle.Parameters.AddWithValue("@d3", CmbBrans.Text);
+                DoktorEkle.Parameters.AddWithValue("@d4", MskTc.Text);
+                DoktorEkle.Parameters.AddWithValue("@d5", TxtSifre.Text);
+                DoktorEkle.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Doktor eklemesi yapıldı", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //doktorları tekrar gridview çek
+                DoktorCek();
+            }
+            else
+            {
+                MessageBox.Show("Doktor eklemek için boş hücre bırakmayınız ", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -71,34 +78,41 @@ namespace HastaneYönetim
 
         private void BtnGuncelle_Click(object sender, EventArgs e)
         {
-            SqlCommand DoktorGuncelle=new SqlCommand("Update Tbl_Doktorlar Set DoktorAd=@d1,DoktorSoyad=@d2,DoktorBrans=@d3,DoktorSifre=@d4 Where DoktorTC=@d5",bgl.baglanti());
-            DoktorGuncelle.Parameters.AddWithValue("@d1",TxtAd.Text);
-            DoktorGuncelle.Parameters.AddWithValue("@d2", TxtSoyad.Text);
-            DoktorGuncelle.Parameters.AddWithValue("@d3", CmbBrans.Text);
-            DoktorGuncelle.Parameters.AddWithValue("@d4", TxtSifre.Text);
-            DoktorGuncelle.Parameters.AddWithValue("@d5", MskTc.Text);
-            DoktorGuncelle.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Bilgiler başarıyla güncellendi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //doktorları tekrar gridview çek
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Doktorlar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+            if (MskTc.Text != "" && MskTc.MaskCompleted && TxtAd.Text != "" && TxtSoyad.Text != "" && CmbBrans.Text != "" && TxtSifre.Text != "") {
+                SqlCommand DoktorGuncelle = new SqlCommand("Update Tbl_Doktorlar Set DoktorAd=@d1,DoktorSoyad=@d2,DoktorBrans=@d3,DoktorSifre=@d4 Where DoktorTC=@d5", bgl.baglanti());
+                DoktorGuncelle.Parameters.AddWithValue("@d1", TxtAd.Text);
+                DoktorGuncelle.Parameters.AddWithValue("@d2", TxtSoyad.Text);
+                DoktorGuncelle.Parameters.AddWithValue("@d3", CmbBrans.Text);
+                DoktorGuncelle.Parameters.AddWithValue("@d4", TxtSifre.Text);
+                DoktorGuncelle.Parameters.AddWithValue("@d5", MskTc.Text);
+                DoktorGuncelle.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Bilgiler başarıyla güncellendi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //doktorları tekrar gridview çek
+                DoktorCek();
+            }
+            else
+            {
+                MessageBox.Show("Bilgileri güncellemek için boş hücre bırakmayınız", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+
         }
 
         private void BtnSil_Click(object sender, EventArgs e)
         {
-            SqlCommand DoktorSil=new SqlCommand("Delete From Tbl_Doktorlar Where DoktorTC=@d1",bgl.baglanti());
-            DoktorSil.Parameters.AddWithValue("@d1", MskTc.Text);
-            DoktorSil.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Kayıt başarıyla silindi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            //doktorları tekrar gridview çek
-            DataTable dt1 = new DataTable();
-            SqlDataAdapter da1 = new SqlDataAdapter("Select * From Tbl_Doktorlar", bgl.baglanti());
-            da1.Fill(dt1);
-            dataGridView1.DataSource = dt1;
+            if (MskTc.Text != "" && MskTc.MaskCompleted && TxtAd.Text != "" && TxtSoyad.Text != "" && CmbBrans.Text != "" && TxtSifre.Text != "") {
+                SqlCommand DoktorSil = new SqlCommand("Delete From Tbl_Doktorlar Where DoktorTC=@d1", bgl.baglanti());
+                DoktorSil.Parameters.AddWithValue("@d1", MskTc.Text);
+                DoktorSil.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Kayıt başarıyla silindi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                //doktorları tekrar gridview çek
+                DoktorCek();
+            }
+            else
+            {
+                MessageBox.Show("Kaydı silmek için seçim yapınız", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void button1_Click(object sender, EventArgs e)
